Add Demoapp framework - not building
