--- OMSDK/OMSDK_Demobuild.framework/Headers
@@ -1,1 +1,0 @@
-Versions/Current/Headers
