--- OMSDK/OMSDK_Demobuild.framework/Modules/module.modulemap
@@ -1,4 +1,0 @@
-module OMSDK_Demobuild {
-    header "Headers/OMIDImports.h"
-    export *
-}
