--- OMSDK/OMSDK_Demobuild.framework/OMSDK_Demobuild
@@ -1,1 +1,0 @@
-Versions/Current/OMSDK_Demobuild
