+++ OMSDK/Service/OMSDK_Demoapp.framework/OMSDK_Demoapp
@@ -1,0 +1,1 @@
+Versions/Current/OMSDK_Demoapp
