+++ OMSDK/Service/OMSDK_Demoapp.framework/Versions/A/Headers/OMIDFriendlyObstructionType.h
@@ -1,0 +1,14 @@
+//
+//  OMIDFriendlyObstructionType.h
+//  AppVerificationLibrary
+//
+//  Created by Andrew Whitcomb on 4/3/19.
+//  Copyright © 2019 Integral Ad Science, Inc. All rights reserved.
+//
+
+typedef NS_ENUM(NSUInteger, OMIDFriendlyObstructionType) {
+    OMIDFriendlyObstructionMediaControls,
+    OMIDFriendlyObstructionCloseAd,
+    OMIDFriendlyObstructionNotVisible,
+    OMIDFriendlyObstructionOther
+};
