+++ OMSDK/Service/OMSDK_Demoapp.framework/Versions/Current
@@ -1,0 +1,1 @@
+A
